Fix: We don't need a `cmp` variable here
The case that the `cmp` variable is zero is excluded, as this would mean
that the node is already in the tree, whereas this function is assumed
NOT to be called in this case.

So we don't need to check for `cmp == 0`. Because this leaves us two
possible cases, we can simply remove the `cmp` variable and embedd the
hash comparison right into the if condition.